requirements.txt
anyio==3.7.0
certifi==2023.5.7
exceptiongroup==1.1.1
fake-useragent==1.1.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
pyproj==3.6.0
python-magic==0.4.27
sniffio==1.3.0
